+++ src/Core/Orders/Services/OrderService.php
@@ -406,7 +406,7 @@ class OrderService extends BaseService implements OrderServiceInterface
     {
         $order = $this->getByHashedId($id);
 
-        if (!empty($data['vat_no'])) {
+        if (! empty($data['vat_no'])) {
             $order->vat_no = $data['vat_no'];
         }