+++ src/Http/Resources/Transactions/TransactionResource.php
@@ -25,6 +25,7 @@ class TransactionResource extends AbstractResource
             'postcode_matched' => (bool) $this->postcode_matched,
             'status' => $this->status,
             'notes' => $this->notes,
+            'created_at' => $this->created_at,
         ];
     }
 }