+++ tests/TestCase.php
@@ -68,7 +68,7 @@ abstract class TestCase extends \Orchestra\Testbench\TestCase
             ],
             'sagepay' => [
                 'vendor' => 'SagePay',
-            ]
+            ],
         ]);
     }