+++ tests/Unit/Shipping/Services/ShippingMethodServiceTest.php
@@ -5,11 +5,12 @@ namespace Tests\Unit\Shipping\Factories;
 use Tests\TestCase;
 use GetCandy\Api\Core\Orders\Models\Order;
 use GetCandy\Api\Core\Baskets\Models\Basket;
+use GetCandy\Api\Core\Countries\Models\Country;
 use GetCandy\Api\Core\Shipping\Models\ShippingZone;
 use GetCandy\Api\Core\Shipping\Models\ShippingPrice;
+use GetCandy\Api\Core\Customers\Models\CustomerGroup;
 use GetCandy\Api\Core\Shipping\Models\ShippingMethod;
 use GetCandy\Api\Core\Shipping\Models\ShippingRegion;
-use GetCandy\Api\Core\Customers\Models\CustomerGroup;
 use GetCandy\Api\Core\Shipping\Services\ShippingMethodService;
 
 /**
@@ -17,10 +18,16 @@ use GetCandy\Api\Core\Shipping\Services\ShippingMethodService;
  */
 class ShippingMethodServiceTest extends TestCase
 {
-
-
     public function test_can_get_correct_shipping_methods()
     {
+        $country = Country::forceCreate([
+            'name' => 'United Kingdom',
+            'region' => 'Europe',
+            'iso_a_2' => 'GB',
+            'iso_a_3' => 'GBR',
+            'iso_numeric' => 826,
+        ]);
+
         $methods = [
             $methodA = $this->createMethod('A'),
             $methodB = $this->createMethod('B'),
@@ -34,10 +41,14 @@ class ShippingMethodServiceTest extends TestCase
             $zoneD = $this->createZone('D'),
         ];
 
+        $zoneA->countries()->attach($country);
         $zoneA->methods()->attach($methodA);
         $zoneB->methods()->attach($methodA);
+        $zoneB->countries()->attach($country);
         $zoneC->methods()->attach($methodB);
+        $zoneC->countries()->attach($country);
         $zoneD->methods()->attach($methodC);
+        $zoneD->countries()->attach($country);
 
         $this->createRegion('AL', $zoneA);
         $this->createRegion('E', $zoneA);
@@ -55,6 +66,7 @@ class ShippingMethodServiceTest extends TestCase
         // So create an order
         $order = Order::forceCreate([
             'currency' => 'GBP',
+            'shipping_country' => 'United Kingdom',
             'shipping_zip' => 'CM6 6TH',
             'basket_id' => $basket->id,
         ]);
@@ -87,11 +99,6 @@ class ShippingMethodServiceTest extends TestCase
         $options = $service->getForOrder($order->encoded_id);
         $this->assertCount(1, $options);
         $this->assertEquals('Shipping Method A', $options->first()->method->attribute('name'));
-
-        // dd($options);
-        // dd('got here');
-
-
     }
 
     private function createPricing($zones)
@@ -114,7 +121,6 @@ class ShippingMethodServiceTest extends TestCase
                 ]);
             }
 
-
             $price = new ShippingPrice;
             $price->rate = 795;
             $price->min_basket = 2500;
@@ -136,7 +142,7 @@ class ShippingMethodServiceTest extends TestCase
     private function createRegion($region, $zone)
     {
         return ShippingRegion::forceCreate([
-            'country_id' => 79,
+            'country_id' => 1,
             'shipping_zone_id' => $zone->id,
             'region' => $region,
             'address_field' => 'postcode',
@@ -152,7 +158,7 @@ class ShippingMethodServiceTest extends TestCase
     }
 
     /**
-     * Creates a shipping method with a suffix in the name
+     * Creates a shipping method with a suffix in the name.
      *
      * @param string $suffix
      * @return array